Make CellAgent Clone
In order to access the traphs from threads, I needed to make it the
cell agent Clone.  Unfortunately, Box<[Ports]> isn’t, so I moved it
back to NalCell, where it belongs anyway.

+++ src/routing_table.rs
@@ -28,7 +28,7 @@ impl RoutingTable {
 			None => Err(RoutingTableError::Index(IndexError::new(index)))
 		}
 	}
-	pub fn set_entry(&mut self, entry: RoutingTableEntry) { self.entries[entry.get_index()] = entry; }
+	pub fn set_entry(&mut self, entry: RoutingTableEntry) { self.entries[entry.get_index() as usize] = entry; }
 }
 impl fmt::Display for RoutingTable {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
